Delete TileController
