--- Assets/Project/Scripts/GamePlayScene/Tile/TileController.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: d73a1a05ef42465093c6040c6c4d3d66
-timeCreated: 1553353423
